Errors are now caught
The output of the eslint command is now written to a temporary file.
Only if the temporary file is at least 5 bytes long, the output will be
used to update the buffer. In other cases, it could be a `null` output
from `jq`.

This change also fixes the issue with the newline character inside a
string being interpreted as a newline character.

Fixes #2